requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langgraph
psycopg2-binary
pandas
python-dotenv
requests
asyncpg
beautifulsoup4